requirements.txt
aioredis==1.3.1
anyascii==0.2.0
arrow==1.1.0
asgiref==3.4.1
async-timeout==3.0.1
attrs==20.3.0
autobahn==21.3.1
Automat==20.2.0
beautifulsoup4==4.9.0
birdsong==1.3.0
bleach==3.3.0
certifi==2020.12.5
cffi==1.14.5
channels==3.0.3
channels-redis==3.2.0
chardet==4.0.0
ciso8601==2.1.3
comments-wagtail-xtd==1.0.1
constantly==15.1.0
cryptography==3.4.7
daphne==3.0.2
Django==3.2.5
django-admin-interface==0.16.3
django-appconf==1.0.4
django-avatar==5.0.0
django-classy-tags==2.0.0
django-cogwheels==0.3
django-colorfield==0.4.1
django-comments-xtd==2.9.1
django-contrib-comments==2.1.0
django-cors-headers==3.7.0
django-debug-toolbar==3.2.1
django-filter==2.4.0
django-flat-responsive==2.0
django-flat-theme==1.1.4
django-ipware==1.1.2
django-js-asset==1.2.2
django-modelcluster==5.1
django-mptt==0.11.0
django-nyt==1.2
django-pagedown==2.2.0
django-pipeline==2.0.6
django-pwa==1.0.10
django-sekizai==2.0.0
django-taggit==1.4.0
django-treebeard==4.5.1
djangorestframework==3.12.4
docutils==0.17.1
draftjs-exporter==2.1.7
et-xmlfile==1.1.0
hiredis==2.0.0
html2text==2020.1.16
html5lib==1.1
hyperlink==21.0.0
idna==2.10
importlib-metadata==4.0.1
incremental==21.3.0
install==1.3.4
jeepney==0.6.0
keyring==23.0.1
l18n==2020.6.1
lxml==4.6.3
Markdown==3.2.2
markdownify==0.9.0
msgpack==1.0.2
openpyxl==3.0.7
packaging==20.9
Pillow==8.2.0
psycopg2-binary==2.8.6
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.20
pyOpenSSL==20.0.1
pyparsing==2.4.7
python-dateutil==2.8.1
pytz==2021.1
requests==2.25.1
SecretStorage==3.3.1
sentry-sdk==1.0.0
service-identity==18.1.0
six==1.15.0
sorl-thumbnail==12.7.0
soupsieve==2.2.1
sqlparse==0.4.1
tablib==3.0.0
telepath==0.2
tqdm==4.60.0
Twisted==21.2.0
txaio==21.2.1
Unidecode==1.2.0
urllib3==1.26.4
wagtail==2.13
wagtail-autocomplete==0.6.3
wagtail-color-panel==1.2.0
wagtail-modelchooser==2.13.1
wagtailfontawesome==1.2.1
wagtailmenus==3.0.2
wagtailnews==2.7.1
wagtailstreamforms==3.15
webencodings==0.5.1
wget==3.2
wiki==0.7.5
Willow==1.4
xlrd==2.0.1
XlsxWriter==1.4.0
xlwt==1.3.0
zipp==3.4.1
zope.interface==5.4.0
